chore: move util classes to util package
